requirements.txt
altgraph==0.17.4
audioop-lts==0.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comtypes==1.4.12
gTTS==2.5.4
idna==3.10
mysql-connector-python==9.4.0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
playsound==1.2.2
PyAudio==0.2.14
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pypiwin32==223
python-dotenv==1.1.1
pyttsx3==2.99
pywin32==311
pywin32-ctypes==0.2.3
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
setuptools==80.9.0
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
